+++ src/runez/__init__.py
@@ -20,7 +20,7 @@ from runez.program import check_pid, is_executable, make_executable, run, which
 from runez.serialize import read_json, represented_json, save_json, Serializable
 from runez.system import abort, decode, expanded, flattened, quoted, short, stringified
 from runez.system import AdaptedProperty, Anchored, CaptureOutput, CurrentFolder, Slotted, TempArgv, TrackedOutput, Undefined, UNSET
-from runez.system import current_test, first_line, get_version, is_tty, PY2, resolved_path, WINDOWS
+from runez.system import first_line, get_version, is_tty, PY2, resolved_path, WINDOWS
 
 __all__ = [
     "DRYRUN",
@@ -42,7 +42,8 @@ __all__ = [
     "read_json", "represented_json", "save_json", "Serializable",
     "abort", "decode", "expanded", "flattened", "quoted", "short", "stringified",
     "AdaptedProperty", "Anchored", "CaptureOutput", "CurrentFolder", "Slotted", "TempArgv", "TrackedOutput", "Undefined", "UNSET",
-    "current_test", "first_line", "get_version", "is_tty", "PY2", "resolved_path", "WINDOWS",
+    "first_line", "get_version", "is_tty", "PY2", "resolved_path", "WINDOWS",
 ]
 
 DRYRUN = False
+color.activate_colors()